Move all utils to the global styles file
